Prove consistency of Imperative.Impl
Give an (erased) implementation of Imperative.Impl in terms of pure
Agda, which shows that the declared interface is consistent. This
is necessary for Imperative.ST to be correct (that Imperative.ST
actually implements Imperative.Impl is not proven).

Create a file for required lemmas that are not in themselves
interesting. Turn on --warning=noUnsupportedIndexedMatch where
applicable because agda-stdlib uses it anyway and Imperative.Pure is not
intended to be used for evaluation.