Lower level of Program
In both Imperative.ST and Imperative.Pure, Program needs only to live in
the same universe as the return value.

Remove the dependence on the levels of the Conditions in Program.
Further simplify by removing the Level parameter on Assignment and the
index on Condition. Consequentially, remove many Level arguments and all
Lifts.