Provide separation proofs to Programs
Extend Imperative.Impl with an operation `separate` that returns a proof
that the references live in the precondition are pairwise distinct.
Extend both Imperative.Impl and Imperative.ST with this operation.

Define a record Erased for wrapping erased values. As Unique is a data
type with constructors that might be matched on, have the new operation
return only an Erased Unique. Add an Unsafe shim for making Eraseds in
Imperative.ST following the example of Unsafe.Realizer.