Add a "smart do notation" for Program
Use the Restructuring solver (still incomplete!) to provide alternate
versions of the Program _>>=_ and _>>_ operators that try to
automatically reorder the program state and use the frame rule to make
preconditions line up. Use this new facility in the Fibonacci example
(which becomes much clearer; success!).

Adjust some comments (esp. removing a note about Cubical and explaining
some of the technical details). Regularize a variable name.