Add Framing
Supply a type Framing that wraps a Restructuring but carries an extra
index to track what is being discarded. Take a Framing in what used to
be reframe (now ManualStyle.frame). This prepares for situations where
evaluation of the underlying Restructuring proof may get stuck, in which
case the version of reframe using discard would get stuck. Design the
"constructors"/introduction rules for Framing such that it should be
easy to produce Framing proofs programmatically by simply following the
tree structure of the inner and outer Conditions.

Reorganize stuff out of Imperative into two new files: AutomaticStyle
and ManualStyle. Export all the names needed to write a program from
each of these modules, and correspondingly simplify imports in the
examples. Put the @tactic-annotated operators into AutomaticStyle and
expose the Framing "constructors" from ManualStyle, defining the two
different ways to handle structural rules while programming.

Flatten out Imperative.Automation again: subsume SmartDo into
AutomaticStyle and put Solvers under Imperative. Add a dumb shim tactic
that solves a Framing goal by just using the existing Restructuring
solver. As above, the plan is to make framing-tactic the center of the
logic and have restructuring-tactic (renamed from restructure!-tactic)
be the simple wrapper. Rename restructure! to restructuring! to avoid
conflating the Program action with the act of coming up with a
Restructuring.

Add a large equality type (in Setω₀) and prove the Condition identity
assoc& in it. Use assoc& to define Framing.

Rethink the type signature of the automatic return. Remove the explicit
postcondition argument since it should indeed be inferrable. Fix the
actual problem encountered before, which is that since the automation
only applies to the _left_ of _>>=_ and _>>_, returns (which usually
appear on the _right_) usually don't get automatically framed. Simply
give return its own @tactic argument to resolve these mismatches.

Revise the Fibonacci examples again to match the newly presented
interfaces.