Make arrays primitive instead of single cells
Dispense with the headache-inducing treatment of Arrays as functions
returning a Ref for each index. In particular, the old treatment seems
to necessitate proofs that appear under a lambda inside the program
state, which defeats unification even though the proof is marked
irrelevant. Also, situations arise where either functional
extensionality is required, or functions needs to be manually proved to
respect function equality. Finally, there is no good (efficient) way to
"increment a pointer in an array", since doing so builds a long chain of
function compositions in memory instead of just updating an integer
index.

Introduce instead the concept of a Slice: an Array always represents a
whole allocation, while a Slice represents some (sub)range of an Array.
Expose the structure of Slices (a record of an array and an index, among
some erased things) to the user, and redefine Ref to be a Slice of
length 1.

The definitions of Slice and some associated functions need to be given
before the definition of Imperative.Impl. Instead of further bloating
Imperative.agda, split off Imperative.Spec as a new file-level module
Imperative.Specifications.