Implement an insertion sort example
Prove functional correctness of the classic in-place array insertion
sort algorithm (for natural numbers): given an array containing numbers,
after the algorithm has run the array now holds a permutation of its old
values and is now sorted. Define what it means for two vectors to be
"Permutation"s, and prove a bunch of lemmas that are used in the
implementation. Use ManualStyle to write the algorithm for now, both to
see how structural rules are used in a less trivial algorithm and also
since the automation is not currently written to handle arrays anyway.

Make a few organizational changes in the library again. Remove
Imperative.Specifications in favor of having Imperative.Condition and
Imperative.Restructuring define their respective data types, as is
traditional. Provide a few utilities here and there that turn out to be
useful in InsertionSort.

Provide (not currently tail-recursive) utilities for reading and writing
whole arrays. Let the level and type of the previous value during a
write be erased (which is required for the new writeArray to work).
Also, require that the postcondition of the argument of runProgram is
just the trivial condition. Note that no Refs of the right type and
scope to appear in the postcondition should ever exist anyway.